Centralise implementation of cancel button logic
